[FIX] replaced asserts with Exceptions
